requirements.txt
pytest-playwright==0.3.3
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1